convert to using MySQL instead of txt files
Change the sensor reader script to write to a MySQL DB instead of a text
file. The MySQL DB can be accessed by an app if a web server is
provided, and it can do all the heavy lifting associated with sorting
data for us. so that we don't need to do it on the app side.

Also add a helper file in which to add all MySQL related functionality.

Change-Id: I48f9acc37174bf767274e37663ced2f4b5b9f4d3
Signed-off-by: Prashant Malani <pmalani@gmail.com>