Connect sensor readings to MySQL server
Clean up the read_init.py sensor code.
Tie it to the main python script.
Make modifications to the DB entry code (we now save the unix timestamp,
which is more portable and less prone to parsing error on the app side)

Make the main python script call the sensor function a few times, in
case the first invocations return errors.

Signed-off-by: Prashant Malani <p.malani@gmail.com>