Handle invalid URLs slightly better
Now, we just don't do anything when an invalid URL is passed.
Eventually we'd like a TOAST message or some text to be displayed, but
for now this is better than a crash.

Also set the Y axis to an appropriate scale. Again this can be done
better, but for the time being this is sufficient.

Signed-off-by: Prashant Malani <p.malani@gmail.com>